+++ src/forge_core/adapters/templates/redshift/create_root_aggregate.sql
@@ -11,7 +11,7 @@ select
     ,'[' || JSON_SERIALIZE(j."root") || ']' as "root"
     ,MD5(JSON_SERIALIZE("root")) as ingestion_hash
     ,CURRENT_TIMESTAMP as ingestion_timestamp 
-    ,'root' as table_path
+    ,'~ROOT_TABLE_PATH~' as table_path
 from 
 (
 ~SQL_SELECTS~